Fix method signatures

+++ src/dates.jl
@@ -113,7 +113,7 @@ end
     end
 end
 
-@inline function Dates.tryparsenext(d::Dates.Delim{<:AbstractChar, N}, str::AbstractVector{UInt8}, i, len) where N
+@inline function Dates.tryparsenext(d::Dates.Delim{<:AbstractChar, N}, str::AbstractVector{UInt8}, i::Int, len) where N
     for j = 1:N
         i > len && return nothing
         next = iterate(str, i)
@@ -124,7 +124,7 @@ end
     return true, i
 end
 
-@inline function Dates.tryparsenext(d::Dates.Delim{String, N}, str::AbstractVector{UInt8}, i, len) where N
+@inline function Dates.tryparsenext(d::Dates.Delim{String, N}, str::AbstractVector{UInt8}, i::Int, len) where N
     i1 = i
     i2 = firstindex(d.d)
     for j = 1:N
@@ -144,7 +144,7 @@ end
     return true, i1
 end
 
-@inline function Dates.tryparsenext_base10(str::AbstractVector{UInt8}, i, len, min_width=1, max_width=0)
+@inline function Dates.tryparsenext_base10(str::AbstractVector{UInt8}, i::Int, len::Int, min_width=1, max_width=0)
     i > len && return nothing
     min_pos = min_width <= 0 ? i : i + min_width - 1
     max_pos = max_width <= 0 ? len : min(i + max_width - 1, len)
